LayoutTests/http/tests/inspector/compiler-source-mapping-provider.html

<html>
<head>
<script src="inspector-test.js"></script>

<script>

function test()
{
    InspectorTest.runTestSuite([
        function testLoad(next)
        {
            var provider = new WebInspector.CompilerSourceMappingProvider("http://localhost:8000/inspector/resources/compiler-source-mapping-provider/app-map.html?sourceMap=app-map");
            var sourceMapping;
            provider.loadSourceMapping(didLoadSourceMapping);
            function didLoadSourceMapping(sourceMappingArg)
            {
                sourceMapping = sourceMappingArg;
                var sources = sourceMapping.sources();
                InspectorTest.assertEquals(2, sources.length);
                InspectorTest.assertEquals("app.js", sources[0]);
                InspectorTest.assertEquals("foo.js", sources[1]);
                provider.loadSourceCode(sourceMapping.sources()[0], didLoadSourceCode);
            }

            function didLoadSourceCode(content)
            {
                InspectorTest.assertEquals(0, content.indexOf("window.onload = function()"));
                provider.loadSourceCode(sourceMapping.sources()[1], didLoadSourceCode2);
            }

            function didLoadSourceCode2(content)
            {
                InspectorTest.assertEquals(0, content.indexOf("function ClickHandler()"));
                next();
            }
        },

        function testTimeout(next)
        {
            var provider = new WebInspector.CompilerSourceMappingProvider("http://localhost:8000/does-not-exist");
            provider.loadSourceMapping(didLoadSourceMapping);
            function didLoadSourceMapping(sourceMapping)
            {
                InspectorTest.assertEquals(null, sourceMapping);
                next();
            }
        }
    ]);
};

</script>

</head>

<body onload="runTest()">
<p>Tests CompilerSourceMappingProvider class.</p>
</body>
</html>
